requirement.txt
asgiref==3.5.2
Django==4.0.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
pandas==1.4.2
numpy==1.21.2